+++ liquid/builtin/tags/unless_tag.py
@@ -1,7 +1,10 @@
 """Tag and node definition for the built-in "unless" tag."""
+from __future__ import annotations
 
 import sys
+
 from typing import TextIO
+from typing import TYPE_CHECKING
 
 from liquid.ast import Node
 from liquid.ast import BlockNode
@@ -10,7 +13,6 @@ from liquid.context import Context
 from liquid.lex import tokenize_boolean_expression
 from liquid.stream import TokenStream
 from liquid.tag import Tag
-from liquid.expression import Expression
 
 from liquid.parse import get_parser
 from liquid.parse import expect
@@ -21,6 +23,9 @@ from liquid.token import TOKEN_EOF
 from liquid.token import TOKEN_EXPRESSION
 from liquid.token import TOKEN_TAG
 
+if TYPE_CHECKING:
+    from liquid import Environment
+    from liquid.expression import Expression
 
 TAG_UNLESS = sys.intern("unless")
 TAG_ENDUNLESS = sys.intern("endunless")
@@ -52,19 +57,25 @@ class UnlessTag(Tag):
     name = TAG_UNLESS
     end = TAG_ENDUNLESS
 
+    def __init__(self, env: Environment):
+        super().__init__(env)
+        self.parser = get_parser(self.env)
+
+    def parse_expression(self, stream: TokenStream) -> Expression:
+        expect(stream, TOKEN_EXPRESSION)
+        expr_iter = tokenize_boolean_expression(stream.current.value)
+        return parse_boolean_expression(TokenStream(expr_iter))
+
     def parse(self, stream: TokenStream) -> UnlessNode:
-        parser = get_parser(self.env)
 
         expect(stream, TOKEN_TAG, value=TAG_UNLESS)
         tok = stream.current
         stream.next_token()
 
-        expect(stream, TOKEN_EXPRESSION)
-        expr_iter = tokenize_boolean_expression(stream.current.value)
-        expr = parse_boolean_expression(TokenStream(expr_iter))
-
+        expr = self.parse_expression(stream)
         stream.next_token()
-        consequence = parser.parse_block(stream, ENDUNLESSBLOCK)
+
+        consequence = self.parser.parse_block(stream, ENDUNLESSBLOCK)
 
         expect(stream, TOKEN_TAG, value=TAG_ENDUNLESS)
         return UnlessNode(tok, condition=expr, consequence=consequence)